refactor(app-suspension): split modules
